refactor: 공통 모킹이 필요한 API를 Super Class로 분리
