+++ client/tests/POM/pom.ts
@@ -1,9 +1,29 @@
 import { Page } from "@playwright/test";
+import { ServerURL } from "constants/serverURL";
+import { VISITOR_MOCK_DATA } from "mocks/data/visitor";
 
 export default class POM {
   readonly page: Page;
 
   constructor(page: Page) {
     this.page = page;
+  }
+
+  async mocking() {
+    await this.page.route(`${ServerURL}/visitor`, async (route) => {
+      const method = route.request().method();
+
+      switch (method) {
+        case "GET":
+          await route.fulfill({
+            json: VISITOR_MOCK_DATA,
+          });
+          return;
+        case "POST":
+          await route.fulfill({
+            json: { todayVisitor: 1, totalVisitor: 10 },
+          });
+      }
+    });
   }
 }
